Add the invoice context & interfaces
Wanted a way to update a property without having to pass it as a prop to children or having to use redux. Decided on using context. Hardest part of using context was, how to update the context from a child component and have that cause a re-render. Ended up following the solution here https://medium.com/digio-australia/using-the-react-usecontext-hook-9f55461c4eae. Which pretty much creates an Provider to pass through a context with a re-defined function to update the property.