Add components for invoice form
This commit attempts to use the invoice context and the react-hook-form to proccess the input fields. The idea is to use React-hook-form in individual components so they are standalone. However, there are clear workflow difficulties as all components will  need to subscibe to a publisher to notify when the data should be submitted. Also, another publisher is required to notify when state for all form components have successfully been registered by react. Committing the code to log and elaborate on learnings from this approach:
- useContext makes it easier to retrieve information from a provider no matter how nested the component is. However, updates to the data will trigger a re-render of every other component which uses the context from that provider. This will make it ineffecient relative to passing props from parents to child and only the child component that uses "said" prop is updated.
- Current usage of react-hook-form is ineffective and creates a messy workflow.
- MaterialUI doesn't keep up to date with the latest react so strict had to be removed from the index.tsx