Add basic pdf template, previewer and downloader
Decided to use react-pdf because it provided a client side solution to generating pdf. It also comes with a previewer so creating & styling pdf was relatively straightforward. Code committed here is a basic and a proof of concept that this solution suits the needs. Will need to refactor, abstract and componentize key features here.
See more documentation on react-pdf below:
https://react-pdf.org/