Add a download pdf button and hidden downloader
So there were a few isues when trying to implement the download pdf feature.
1. PDFDownloadLink component does not have an accessible callback or event to indicate whether the pdf has completing loading
2. The pdf will only download when PDFDownloadLink has processed the pdf and is clicked

To address the first problem. A child component LoadedPDFButton with a callback was created.  PDFDownloadLink is a HOC which will pass a loading property to the child (LoadedPDFButton). A callback in the child component will fire upon each loading change.

To address the second problem a HiddenInvoiceDownloader component was created. This component is not visible to the DOM. This component has a ref to the parent div wrapping the PDFDownloadLink. A callback  to click the PDFDownloadLink when loading has finished, is passed to the LoadedPDFButton.