Refactor UserContext to AppContext & rename action
UserContext does not make sense. The aim of the context is to be a sourceof truth for globally accessible data - so not neccessarily data related to user context. Hence the rename to AppContext

Also remaned user action types to be {ACTION}_USER_DETAILS. Also adde a generic action type for deleting