Refactor sign-in tests & remove ts-jest
Tried doing class mocks from import following
https://jestjs.io/docs/en/es6-class-mocks
however, there were difficulties because we have multiple exports in the one firebase.context file so we would have to mock all the other ones (can work around this  by hijacking the export default). Another issue with this is that the mock is defined at the top on the class. This class will be used in other tests which would require different mock return data. This would make the testing process tedious

Sticking with the current implementation(minus a few unneeded lines of code and packages). Current implementation relies on all exports being undefined and will then use prototypic assignment on the Firebase class to redefine the function inline. The firebase object will need to be re-instantiated with each re-definition. Mock data does not need to be cleared because service is re-instantiated each time. This is a bit hacky because even the FirebaseContext is undefined, yet it is still able to work as a react context.