Re-factor sign-in tests to use jest.spyOn
After doing more research found a better way to mock data. The below link has a comment at the bottom with a spyOn example
https://dev.to/jackcaldwell/mocking-es6-class-methods-with-jest-bd7
reading this comment lead me to look at the spyOn docummentation.
https://jestjs.io/docs/en/jest-object
SpyOn is applied to the function of an object. It works by tracking function calls to it and will override the original function with the mocked implementation. Original functionality can be restored with jest.restoreAllMocks(). This is methodof unit testing is more advantageous to the previous method because we:
1. Don't have to keep track of mockImplementation, as there is a built-in clean up method
2. Only mock what is neccessary and maintain the integrity of the imports (previous method made them all undefined)
3. Can easily perform integration testing with original functions
Disadvantage to this is that Firebase SDK need to be initialised with each test-suite