Add alert system (context, reducer, ui, test)
Separted the alert context because alerts will only be associated with the alert related components and UI. So created it based on separate responsibility

Implemented custom transition animation for alert card because material UI Slide component would not display animation beccause it would render mount child elements on show & if the child element is re-rendered (because of passing state) then the transition is skipped.

Test cases for alerts were implemented differently to other tests. Will be doing more of the following:
- Using screen object when interacting with DOM. This is more convenient then destructoring the render
- Not using afterEach(cleanup), because this is already implemented by the framework
- using findBy* queries instead of waitFor*s. findBy* does a waitFor in the background, but findBy is simpler and conveniet
- Limit using *ByTestIds and using ByRole* and ByLabel* etc more to reflect user interaction more (it has more methods that resemble user events)
See the below link for more information
https://kentcdodds.com/blog/common-mistakes-with-react-testing-library