Add sidebar for nav and remove links on navbar
Using sidebar for navigation to reduce clutter on navbar. There will be a lot of route options and having it all on navbar will be messy.

Sidebar will only appear if the user has signed in.

Added a Collapsible and Permanent Sidebar. Each will render on different screen sizes.

Moving forward I will be following the naming convention as stated in
https://medium.com/hackernoon/structuring-projects-and-naming-components-in-react-1261b6e18d76