Add dateUpdated to inv & submit flag to Inv form
DateUpdated so update time is tracked. Also use it to trigger re-download of pdf as the hidden invoice download needs a new object to re-trigger.

Add hasSubmitted flag to invoice form. This makes it clear when the user has submitted/saved the invoice.

Also add a check to see if the submittedInvoice matches the previous invoice. This will limit writes to the db if the data is the same.